first part of record form with the record creation
