+++ models/Resident.ts
@@ -1,19 +1,16 @@
 import mongoose, { Document, Schema } from "mongoose";
 import { IUser } from "./User";
-import { ITeacher } from "./Teacher";
 
 export interface IResident extends Document {
     _id: mongoose.Types.ObjectId;
     user: IUser
-    teachers: ITeacher[];
     createdAt: Date;
     updatedAt: Date;
 }
 
 const ResidentSchema = new Schema<IResident>(
     {
-      user: { type: Schema.Types.ObjectId, ref: "User", required: true },
-      teachers: { type: [{ type: Schema.Types.ObjectId, ref: "Teacher" }], default: [] },
+      user: { type: mongoose.Types.ObjectId, ref: "User", required: true },
     },
     { timestamps: true }
 );
